# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Pa Emmanuel Ayodele Abatan memorial invitation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  id: number;
  deceasedName: string;
  eventTitle: string;
  waykeepDate: string;
  burialDate: string;
  /** @nullable */
  year?: number | null;
  venue: string;
  dressCode: string;
  /** @nullable */
  biography?: string | null;
  /** @nullable */
  tribute?: string | null;
  /** @nullable */
  importantInformation?: string | null;
  /** @nullable */
  directions?: string | null;
  /** @nullable */
  contactInformation?: string | null;
  /** @nullable */
  mapUrl?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  backgroundImageUrl?: string | null;
}

export interface EventUpdate {
  deceasedName?: string;
  eventTitle?: string;
  waykeepDate?: string;
  burialDate?: string;
  /** @nullable */
  year?: number | null;
  venue?: string;
  dressCode?: string;
  /** @nullable */
  biography?: string | null;
  /** @nullable */
  tribute?: string | null;
  /** @nullable */
  importantInformation?: string | null;
  /** @nullable */
  directions?: string | null;
  /** @nullable */
  contactInformation?: string | null;
  /** @nullable */
  mapUrl?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  backgroundImageUrl?: string | null;
}

export interface ProgrammeItem {
  id: number;
  title: string;
  date: string;
  /** @nullable */
  time?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder: number;
}

export interface ProgrammeItemInput {
  title: string;
  date: string;
  /** @nullable */
  time?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
}

export interface ProgrammeItemUpdate {
  title?: string;
  date?: string;
  /** @nullable */
  time?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
}

export type PublicInvitationStatus = typeof PublicInvitationStatus[keyof typeof PublicInvitationStatus];


export const PublicInvitationStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  checked_in: 'checked_in',
  disabled: 'disabled',
} as const;

export type PublicInvitationRsvpStatus = typeof PublicInvitationRsvpStatus[keyof typeof PublicInvitationRsvpStatus];


export const PublicInvitationRsvpStatus = {
  pending: 'pending',
  yes: 'yes',
  no: 'no',
} as const;

export interface PublicInvitation {
  guestName: string;
  invitationCode: string;
  admissionLimit: number;
  status: PublicInvitationStatus;
  rsvpStatus: PublicInvitationRsvpStatus;
  event: Event;
  programme: ProgrammeItem[];
  /** @nullable */
  checkedInAt?: string | null;
}

export type RsvpInputResponse = typeof RsvpInputResponse[keyof typeof RsvpInputResponse];


export const RsvpInputResponse = {
  yes: 'yes',
  no: 'no',
} as const;

export interface RsvpInput {
  response: RsvpInputResponse;
}

export interface GuestInput {
  /** @minLength 2 */
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @minimum 1 */
  admissionLimit?: number;
}

export interface GuestUpdate {
  /** @minLength 2 */
  fullName?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @minimum 1 */
  admissionLimit?: number;
}

export type AdminGuestStatus = typeof AdminGuestStatus[keyof typeof AdminGuestStatus];


export const AdminGuestStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  checked_in: 'checked_in',
  disabled: 'disabled',
} as const;

export type AdminGuestRsvpStatus = typeof AdminGuestRsvpStatus[keyof typeof AdminGuestRsvpStatus];


export const AdminGuestRsvpStatus = {
  pending: 'pending',
  yes: 'yes',
  no: 'no',
} as const;

export interface AdminGuest {
  id: number;
  invitationId: number;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  invitationCode: string;
  invitationToken: string;
  admissionLimit: number;
  status: AdminGuestStatus;
  rsvpStatus: AdminGuestRsvpStatus;
  createdAt: string;
  /** @nullable */
  checkedInAt?: string | null;
  /** @nullable */
  checkedInBy?: string | null;
}

export interface GuestImportInput {
  /** @minLength 1 */
  csv: string;
}

export type GuestImportResultErrorsItem = {
  row: number;
  message: string;
};

export interface GuestImportResult {
  created: AdminGuest[];
  errors: GuestImportResultErrorsItem[];
}

export interface InvitationLookupInput {
  /** @minLength 4 */
  code: string;
}

export type InvitationLookupResultResult = typeof InvitationLookupResultResult[keyof typeof InvitationLookupResultResult];


export const InvitationLookupResultResult = {
  valid: 'valid',
  used: 'used',
  disabled: 'disabled',
  invalid: 'invalid',
} as const;

export interface InvitationLookupResult {
  result: InvitationLookupResultResult;
  invitation?: AdminGuest | null;
}

export type AdmissionResultResult = typeof AdmissionResultResult[keyof typeof AdmissionResultResult];


export const AdmissionResultResult = {
  admitted: 'admitted',
  used: 'used',
  disabled: 'disabled',
  not_found: 'not_found',
} as const;

export interface AdmissionResult {
  result: AdmissionResultResult;
  invitation?: AdminGuest | null;
  /** @nullable */
  checkedInAt?: string | null;
}

export interface AuditRecord {
  id: number;
  action: string;
  /** @nullable */
  target?: string | null;
  actor: string;
  createdAt: string;
}

export interface DashboardSummary {
  totalInvited: number;
  confirmed: number;
  pending: number;
  checkedIn: number;
  notCheckedIn: number;
  disabled: number;
  rsvpYes: number;
  rsvpNo: number;
  recentActivity?: AuditRecord[];
}

export interface CheckInRecord {
  id: number;
  guestName: string;
  invitationCode: string;
  checkedInAt: string;
  checkedInBy: string;
}

export interface UploadRequest {
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}

export type UnauthorizedResponse = {
  error?: string;
};

export type NotFoundResponse = {
  error?: string;
};

export type BadRequestResponse = {
  error?: string;
};

export type ListGuestsParams = {
search?: string;
status?: ListGuestsStatus;
rsvp?: ListGuestsRsvp;
};

export type ListGuestsStatus = typeof ListGuestsStatus[keyof typeof ListGuestsStatus];


export const ListGuestsStatus = {
  all: 'all',
  pending: 'pending',
  confirmed: 'confirmed',
  checked_in: 'checked_in',
  disabled: 'disabled',
} as const;

export type ListGuestsRsvp = typeof ListGuestsRsvp[keyof typeof ListGuestsRsvp];


export const ListGuestsRsvp = {
  all: 'all',
  yes: 'yes',
  no: 'no',
  pending: 'pending',
} as const;

export type ListCheckInsParams = {
search?: string;
filter?: ListCheckInsFilter;
};

export type ListCheckInsFilter = typeof ListCheckInsFilter[keyof typeof ListCheckInsFilter];


export const ListCheckInsFilter = {
  all: 'all',
  checked_in: 'checked_in',
  not_checked_in: 'not_checked_in',
  disabled: 'disabled',
  rsvp_yes: 'rsvp_yes',
  rsvp_no: 'rsvp_no',
} as const;