# lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Pa Emmanuel Ayodele Abatan memorial invitation API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadRequest {
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: string;
}